+++ games/static/js/add_session.js
@@ -8,6 +8,9 @@ for (let button of document.querySelectorAll("[data-target]")) {
     event.preventDefault();
     if (type == "now") {
       targetElement.value = toISOUTCString(new Date);
+    } else if (type == "copy") {
+      const oppositeName = targetElement.name == "timestamp_start" ? "timestamp_end" : "timestamp_start";
+      document.querySelector(`[name='${oppositeName}']`).value = targetElement.value;
     } else if (type == "toggle") {
       if (targetElement.type == "datetime-local") targetElement.type = "text";
       else targetElement.type = "datetime-local";