+++ games/views.py
@@ -64,7 +64,7 @@ def edit_session(request, session_id=None):
         return redirect("list_sessions")
     context["title"] = "Edit Session"
     context["form"] = form
-    return render(request, "add.html", context)
+    return render(request, "add_session.html", context)
 
 
 def edit_purchase(request, purchase_id=None):